# openapi/service.v1.yaml
openapi: 3.1.0
info:
  title: sysmlv2-ls-service API
  version: 0.1.0
  description: HTTP API for the SysML v2 / KerML validation microservice.
  license:
    name: Unspecified
security: []
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /v1/validate:
    post:
      operationId: validateSysML
      summary: Validate one request-local SysML/KerML workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRequest"
            examples:
              simple:
                summary: Single SysML document.
                value:
                  files:
                    - uri: memory:///demo.sysml
                      text: package Demo { part def Vehicle; }
      responses:
        "200":
          description: Validation diagnostics and per-file summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateResponse"
        "400":
          description: Request schema validation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: HTTP payload too large.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Validation timeout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    SysMLFileInput:
      type: object
      additionalProperties: false
      required:
        - text
      properties:
        uri:
          type: string
          minLength: 1
          description: Request-local document URI.
        path:
          type: string
          minLength: 1
          description: Request-local file path converted to a document URI.
        language:
          type: string
          enum: [sysml, kerml]
        text:
          type: string
    ValidateRequest:
      type: object
      additionalProperties: false
      required:
        - files
      properties:
        files:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SysMLFileInput"
        standardLibrary:
          type: string
          enum: [none]
          default: none
        validationChecks:
          type: string
          enum: [all, none]
          default: all
    SourcePosition:
      type: object
      additionalProperties: false
      required: [line, character]
      properties:
        line:
          type: integer
          minimum: 0
        character:
          type: integer
          minimum: 0
    SourceRange:
      type: object
      additionalProperties: false
      required: [start, end]
      properties:
        start:
          $ref: "#/components/schemas/SourcePosition"
        end:
          $ref: "#/components/schemas/SourcePosition"
    ServiceDiagnostic:
      type: object
      additionalProperties: false
      required: [severity, source, message, uri, range]
      properties:
        severity:
          type: string
          enum: [error, warning, information, hint]
        source:
          type: string
        code:
          oneOf:
            - type: string
            - type: number
        message:
          type: string
        uri:
          type: string
        range:
          $ref: "#/components/schemas/SourceRange"
    FileValidationSummary:
      type: object
      additionalProperties: false
      required: [uri, language, parserErrors, lexerErrors, diagnostics]
      properties:
        uri:
          type: string
        language:
          type: string
          enum: [sysml, kerml]
        parserErrors:
          type: integer
          minimum: 0
        lexerErrors:
          type: integer
          minimum: 0
        diagnostics:
          type: integer
          minimum: 0
    ValidateMeta:
      type: object
      additionalProperties: false
      required: [standardLibrary, validationChecks, elapsedMs]
      properties:
        standardLibrary:
          type: string
          enum: [none]
        validationChecks:
          type: string
          enum: [all, none]
        elapsedMs:
          type: number
          minimum: 0
    ValidateResponse:
      type: object
      additionalProperties: false
      required: [ok, diagnostics, files, meta]
      properties:
        ok:
          type: boolean
        diagnostics:
          type: array
          items:
            $ref: "#/components/schemas/ServiceDiagnostic"
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileValidationSummary"
        meta:
          $ref: "#/components/schemas/ValidateMeta"
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        issues:
          type: array
          items:
            type: object
            additionalProperties: true